Add subnavbar to layout with dynamic title
