+++ src/App.css
@@ -45,4 +45,5 @@
 
 .pageContainer{
   margin-top: 85px;
+  height: calc(100vh - 85px);
 }
