+++ src/Layouts/Authenticated.js
@@ -1,21 +1,15 @@
 import React from 'react';
 import Navbar from '../Components/Navbar/Navbar';
+import SubNavbar from '../Components/Navbar/SubNavbar';
 
-const menu = [
-  { text: 'Home', icon: 'home', link: '/' },
-  { text: 'Dashboard', icon: 'dashboard', link: '/dashboard' },
-  { text: 'My Class', icon: 'school', link: '/classroom' },
-  { text: 'Forum', icon: 'forum', link: '/topics' },
-  { text: 'Daily', icon: 'today', link: '/daily' },
-];
-
-export default function Authenticated({ children }) {
+export default function Authenticated({ children, title }) {
   return (
     <div className='flex_ lightgreyBackground'>
       <div style={{ width: 200 }}>
-        <Navbar menu={menu} />
+        <Navbar />
       </div>
       <div className='lightgreyBackground' style={{ width: 'calc(100% - 200px)' }}>
+        <SubNavbar title={title} />
         {children}
       </div>
     </div>
